+++ lib/pages/journal/infinite_journal_page.dart
@@ -71,11 +71,13 @@ class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
     super.initState();
   }
 
+  // TODO: move all this to the new JournalPageCubit
   Future<void> _fetchPage(int pageKey) async {
     final cubit = context.read<JournalPageCubit>();
 
     try {
       Set<String>? entryIds;
+      // TODO: rethink tags
       for (final tagId in tagIds) {
         final entryIdsForTag = (await _db.entryIdsByTagId(tagId)).toSet();
         if (entryIds == null) {
@@ -90,10 +92,15 @@ class _InfiniteJournalPageState extends State<InfiniteJournalPage> {
           .whereType<String>()
           .toList();
 
+      final fullTextMatches = cubit.state.fullTextMatches.toList();
+      final ids = fullTextMatches.isNotEmpty ? fullTextMatches : null;
+
       final newItems = await _db
           .watchJournalEntities(
             types: types,
-            ids: entryIds?.toList(),
+            // TODO: bring back tags matching
+            // ids: entryIds?.toList(),
+            ids: ids,
             starredStatuses:
                 cubit.state.starredEntriesOnly ? [true] : [true, false],
             privateStatuses: